packaging: updating packaging script to support Buck binary output
Summary:

With the introduction and support of the Buck build system, we should
start adding support to our various build|provision|deployment scripts
for the Buck components. This adds in checks for the buck binaries as a
secondary option to the CMake binary output to make generation of
osquery packages easy for either build platform.

Test Plan:

CMake:

- Build the binaries:

```

```

- Package the binaries:
```

```

- Install the binaries:
```

```

Buck:

- Build the binaries:
```

```

- Package the binaries:
```

```

- Install the binaries:
```

```