Change MSI Service Error handling (#5467)
The Wix configuration sets the service as _critical_. As I understand it, this means failures to start osquery, are considered startup failures, and will trigger a reboot. As there are occasional windows bugs causing a failure to start, this seems unfortunate. I think hit someone on slack today 

This PR changes that to _normal_

Docs are http://wixtoolset.org/documentation/manual/v3/xsd/wix/serviceinstall.html