deploy: update windows choco scripts to migrate service to Program Files
Summary: This updates the chocolatey deployment scripts to ensure that
if the current installation is using the ProgramData path in the system
service, this is removed and subsequently updated to the new Program
Files location.

Test Plan: Will update with builds and upgrades of the chocolatey
package from 3.3.2 to master.