deploy: updating msi build script (#5794)
Summary: This adds a few minor updates to the MSI build function. First
it keeps build artifacts located in `build/msi` for ease of access.
Further it updates component referneces to be x64 compatible. Finally it
updates a few path references to reflect the new repo layout, as well as
changes a few references for author and ownership information to be
osquery.

Test Plan: Will update with sample MSI builds for Cmake.