Fix MSI Service Error handling
When ErrorControl is set to `critical`, a failure to start osquery results in a system reboot. Instead, this should be set to `normal` where it is logged and the startup proceeds.

Upstream docs are Docs are http://wixtoolset.org/documentation/manual/v3/xsd/wix/serviceinstall.html

This was fixed in https://github.com/osquery/osquery/pull/5467 and seems to have gotten lost.