http: Move away from deprecated async_resolve overload (#5947)
When testing on IPv6-only hosts, with an IPv4 lo stack, the resolver may
not return an IPv4 address. This seems to be an issue with the
deprecated overload of async_resolve. Upon inspection, the previous
overload would only return 1 result_type.

Migrating to the recommended overload works as expected.