macOS pkgbuild and productbuild
## Background

As I understand mac packaging, there are 2 kinds.  packages (build by pkgbuild) and distribution packages (build by productbuild). These both use `.pkg` as the extension and are almost always interchangeable. Distribution packages are somewhat preferred, as they are needed for distribution using an MDM. I do not believe there are any places a simple package is preferred.

Generally speaking, building these is straight forward. Use `pkgbuild` to build the packages, and `productbuild` to convert to distribution packages. cpack seems to have implemented this.

`productbuild` has a bunch of logic around exactly what's it's doing based on the distribution xml. However, it's all optional. At the simplest level, you can use productbuild to convert a package. cpack _does_ use distribution xml. Presumably it's correct.

We have a current issue where we need to change the identifier cpack builds with. It auto assigns one, but we want to keep the legacy facebook one. So, we need to strip off the `--identifier` argument passed from cpack, and replace it.

## Problem

In https://github.com/osquery/osquery/pull/5934 we introduced some stripping logic. However, this script also added some other changes.

It removed the call to `productbuild`. While not always needed, I think this is a bug. Looking at the cpack code, I think they're call to it is fine.

This script also only recognizes some of the options cpack passes in. If and when we add product signing, that would cause failure.

## Proposal

First, let cpack do it's `productbuild` thing. If we find we need to change the `productbuild` invocation, we can add a specific shim there.

Second, be more narrow in how we change the passed arguments. Instead of trying to whitelist the arguments we want, we can remove the undesired one.

## Test Plan

To be honest, I'm not sure. I've never tried to build the osx package. Build and examine the output.