Force git to create symlinks on the CI
git on Windows supports cloning a repository with symlinks
without actually creating them; it will substitute them
with text files containing the target path.
This can potentially break the build of some of osquery
third party libraries.

Moreover this is the default on the CI,
which is different compared to the setup on local machines,
which means that it can hide other kind of issues.