core/{tests,windows}: Refactor WMI method calling
Moves ExecMethod to the WmiRequest class, avoiding the need
for shared_ptr entirely. This is also slightly conceptually cleaner,
since WmiRequest maintains the state for our requests and its scope
should dictate the validity of calling methods on the results it
returns.

Also turns wmiObjectDeleter into a functor, since C++11 lambdas aren't
CopyAssignable and the move assignment operator for WmiResultItem needs
that.