Fix codegen template for extension group
Fix codegen template for extension group. The template has a reference
of `waitThenShutdown` which does not exist in the definitions. Instead
of that, a function, `waitForShutdown` is defined in the
`osquery/system.h` which should have been used. This causes compilation
to fail in case of building an extension. This commit fixes that
function call in the template.