Fix `cmdline` not populated in MacOS
Fix the missing/improper `cmdline` field in the processes table, as
mentioned in #6330. The contents of the call stack was initially read
from the stack pointer, `argc`. That has been moved to start now from

```
char* ptr = procargs + sizeof(nargs);
```

From experiments, it looks like there are leading null-byte chars after
argc[0] that causes the counter `i` to advance thus missing some args
from the `nargs` count. To avoid that, a null-byte sink-hole has been
written that will convert all null-bytes to space by using while instead
of if in the second conditional in the for loop.

```
while (*ptr == '\0' && ptr < procargs + len && i < nargs) {
  *ptr = ' ';
  ptr += 1;
}
```

The last byte is not converted to a space as, argc[1] is also
representing the executable name. So, the parsing loop will only run to
second last argument and then rely on the next null-byte for
termination.

ref:
* https://github.com/skaht/Csu-85/blob/master/start.s
* https://github.com/hishamhm/htop/blob/master/darwin/Platform.c
* Code for experiment: https://paste.ubuntu.com/p/QxTyvZ7cXn/