Update winbaseobj.table
Add common windows object types to winbaseobj table description.  The reasoning behind the change is that what people are looking for is typically an object type such as a mutex.  It is not common knowledge that mutexes, for example, are a type of Windows object.  So when searching to see if osquery supports enumerating mutexs, it is not easy to determine that the answer is yes.