Add target process information on a KILL syscall
The audit messages for KILL contains a message of the type OBJ_PID that
contains information about the process on which the KILL syscall was
issued to.

Add fields ocomm, oauid and oses to the AuditEvent.

Note: This will not be published in the SQL tables as the moment because
the schema isn't modified to accomodate it.