Update hash cache inode number
When the executable has changed on disk and the cache is updated with the new hash value, it was not updating the inode number in the cache used for future lookups. This resulted in the cache being perpetually stale and recomputing the hash for every future lookup of the file.