CI: Python installation improvements on Windows
Since the CI already selects the newest installation of Python,
to then install some additional packages, lets pass the root folder to CMake,
so it doesn't have to detect Python again and possibly select the wrong version.

Remove the need to use Python3 32bit, since it's not necessary
and the Windows installation used is a 64bit one anyway.

Fix pip complaining about the Scripts folder not being in the PATH,
by prepending the PATH with the selected Python installation
root folder and Scripts folder.

Split and reorder the various prerequisite install steps
so that they are closer to each other
and slightly better categorized.