CMake: Detect missing headers during include namespace generation
When a header specified to be added in a generated include namespace
does not exists because its path is wrong or it shoudn't be added,
issue a fatal error, so that it gets noticed and fixed.

Remove windows/userassist.h as a header to be added in an include
namespace, since it doesn't exists.