CMake: Add -pthread compile option on posix platforms
osquery was already linking with -pthread and so linking
to libpthread on Linux, but it wasn't always defining
the _REENTRANT macro which is done by the -pthread option
given at compile time.
Although in the third party libraries that need it,
it should've been defined in other ways (directly or via a config.h),
always add -pthread for correctness and consistency.

Note: macOS doesn't need -pthread at link time because
pthreads are already implemented inside the libc library.