Move try/catch to all partial failure during /proc iteration
As described in #5709, some permissions schemes do not have full
permission on /proc. Move the try/catch block into the loop, and allow
iteration despite errors.

Track whether we've had any success and base the failure off of it.

fixes: #5709

I'm not convinced this is quite right, but it seems worth review