Fix StartupItemTest failing due to unexpected values
The "path" column for a systemd unit row can be empty.

Added missing possible values for the "status" column,
when the "type" column value is "systemd unit".

Removed "Login Item" as a possible value for the "type" column,
since now "Startup Item" is used.

Removed "disabled" as a possible value for the "status" column,
since it's not returned anymore and due to the type "Startup Item"
being either a script or a .desktop file, which do not have a disabled state;
if they need to be they'll just be removed
and they won't appear in the table anymore.

Separately check for the "status" column possible values
since they differ based on the "type" column value.