Augeas Table Updates
While playing with augeas, I noticed it loaded the lenses that were
already on my machine (at `/usr/local/share/augeas/lenses`). This was an
unwelcome surprise.

I'm not sure if this was previously intended or not. I assume not, since
there is some versioning with the lenses, and we ship our own.
This adds the load flags to ignore anything on the existing system path.
And disable saving while I'm here.

Additionally, this adds support for querying against the augeas node, by
both LIKE and EQUALS. As well as bringing in LIKE support for the path
column.

Lastly, it adds some code comments with my understanding of how this
works,

Docs: https://github.com/hercules-team/augeas/blob/master/src/augeas.h

Fixes: #6974