Augeas Table: Don't autoload system lenses
While playing with augeas, I noticed it loaded the lenses that were
already on my machine (at /usr/local/share/augeas/lenses). This was an
unwelcome surprise.

I'm not sure if this was previously intended or not. I assume not, since
there is some versioning with the lenses, and we ship our own.
This adds the load flags to ignore anything on the existing system path.
And disable saving while I'm here.

Note that this is, potentially, an API change. Though I think it's
adding consistency

Fixes: #6974