Update dockerfile to multiplatform
This updates the dockerfile use to make the ci builders to use dockerx,
multistage, and multiplatform builds. There is a Makefile as a
convenience for the command.

Due to the manual cmake build, this is quite slow, and not suitable for
automation.

Note that until we update .github/workflows/build.yml these changes will not take effect

beta builder

Update .github/workflows/build.yml