Consolidate syscalls to a single audit rule
When inspecting the syscalls via the linux audit framework, it's
recommended to have as trimmed down rules as possible. osquery
originally installed individual syscall watches without any custom
filters for each syscall. That can lead to a lot of implicit filtering
within kauditd since it needs to match an event through upto 35 syscalls
individually.

Libaudit allows adding multiple syscalls to be added in the same audit
rule data. This can lead to matching all of the syscalls in a single
match within the kernel. For the record, an audit rule contains an array
that contains information about what syscalls it needs to match within a
bitmap. Which means, multiple syscalls can be matches efficiently using
bitmasks. In practice this leads to much better cpu consumption.

For the audit rules of the form
```
-s always,exit -S execve
-s always,exit -S execveat
```
The new rule will be
```
-s always,exit -S execve,execveat
```