CI: Regenerate sccache cache when compiler version changes
Add the compiler version to the sccache key so that if the version
changes, the key won't match and a new cache key is generated.

Without this all the first runs of PRs after the compiler version
changes, provided they succeed, or until a commit lands on master and
it's built before the PR,
will have sccache miss everything and the build will take double the time.

With this, if a commit lands on master and its build or the scheduled build
run before a new opened PR, then that PR will find the new cache key
and it won't take double the time.