windows: update packaging icon to be png
Summary:

Seems that the cpack builds are not interested in using a `.ico` file, and preferring that we ship a jpeg or PNG or the like.

Here's the failed build: https://github.com/osquery/osquery-codesign/runs/2840361169?check_suite_focus=true
Here's the nuspec for the error: https://docs.microsoft.com/en-us/nuget/reference/errors-and-warnings/nu5046

Test Plan:

Will update with sample builds from the osquery-packaging repo.