Do not let osquery create multiple copies of the extension running at once (#7178)
With the existing behavior, osquery can launch multiple extensions if the running extension uses
a lot of CPU & RAM.

With the new behavior, osquery will never launch multiple extensions.
- if the extension watchdog is enabled, it will kill the old extension then relaunch a new one
- if the extension watchdog is not enabled, it will do nothing

osquery was ignoring the return of `extension.second->checkStatus` and instead looking at
`extension.second->isValid`. `isValid` is just "did the extension ever launch". With the change in
behavior this means that if the extension had been launched once, it would never be relaunched

Co-Authored-By: Ivan Leichtling <ivanlei@apple.com>