Small changes to the `disk_encryption` and `block_devices` tables
While trying to understand some LVM things (which I incorrectly thought
was a bug, see #7207) I noticed that these tables were very slow to work
with. So, a couple of small, speculative, changes.

First, flag them as cachable. I'm not sure this functionality works, but
I think this is more correct.

Second, allow specification of an explicit name to `disk_encryption`.
This implementation _omits_ the uuid that would be inherited from
block_devices, but does allow other names to be queried. It's also
faster.