Fix osquery_info build_platform column value on Linux
The value of build_platform in the osquery_info table
is currently controlled by the value of the preprocessor macro
OSQUERY_BUILD_PLATFORM, which is hardcoded to linux as a macro
passed via CLI.
That macro is then stringified, but the issue is that the linux
keyword is also a macro, which gets expanded to 1.

To avoid this collisions, we can set the preprocessor macro value
as a string already from CLI and remove the stringification,
which will prevent further problematic expansions.

We do the same also for the other string macro OSQUERY_BUILD_DISTRO.