Fix audit rule removal upon osquery exit (#7221)
The audit rules were not getting removed when osquery exits as part of
the cleanup process. This was due to not adding the audit rule to the
cleanup list unless osquery was run with --audit_force_unconfigure. The
force unconfigure option is to remove the rules irrespective of the their
install exit code which was not followed in the #7063.

The current fix is going to install the rule for cleanup either the error
code is non-negative or osquery is asked to force reconfigure. This would
also fix the double adding of the audit rules as mentioned in #7205.