libs: Add M1 support to ssdeep and update it to the latest commit
Reduced what the library was compiling; the C++ parts were unused,
and were parts of the ssdeep application.

The library has been updated to latest master (d8705da),
since the latest release is from 2017 and the library
last received an update in 2019.

Update the README to remove information which is now aggregated under
libraries/cmake/README.md