cmake: Do not compile smartmontools and table on M1
smartmontools source code doesn't provide a library;
the one we have is custom built on top of the original code,
via several changes.
This library is not anymore supporting, not easy to reproduce/update
to use another version of smartmontools and it only supports
the SATA protocol, which is not used by newer macbook hardware.

Fixing this is out of scope.