libs: Add M1 support to glog version 0.5.0
Define a custom prefix for log lines, which maintains the format as
the old glog version.

Update the README to remove information which is now aggregated under
libraries/cmake/README.md

Update the order of the config folders so that it's os first,
architecture after, given that if the library config has something
in common, it will be across architectures in the same os.