Fix parental encryption status with EQUALS constraint
https://github.com/osquery/osquery/issues/7314

An EQUALS constraint is applied explicitly by the user with a WHERE in the
query, and implicitly by a JOIN. When such an EQUALS constraint is applied to
the disk_encryption virtual table, parent devices also needs to be queried
recursively. Otherwise, a device that is the child of an encrypted device will
incorrectly report as unencrypted when queried with WHERE or JOIN, but will
correctly report as encrypted when queried without such a constraint.