cmake: Add support to natively compile and cross-compile for M1
Permit to compile one OSX architecture at a time.
Since there are different sets of source files to build,
the configuration needs to be run twice.

Add an alias of the "arm64" macOS system processor to "aarch64"
to recognize native compilation on M1.

Add support to cross-compilation from x86_64 to arm64 for macOS,
when the target architecture is in CMAKE_OSX_ARCHITECTURES.