Implement a performant cache for users and groups on Windows
Add two new services UsersService and GroupsService,
which will scan and cache users and groups information in memory.

The speed of the scan is configurable with two new flags,
users_service_delay and groups_service_delay.
The services will scan 100 users or groups at a time
and then use the above delay to throttle.
The interval between full scans is configurable via other two new flags,
users_service_interval and groups_service_interval.

While building the cache an optimized indexing is created
for the columns that are marked as index.

The very first time it's run, the services will first do a full scan
before providing any results, so any table or other code
that wants to access the cache will have to wait.
After the cache is initialized, the cache is updated incrementally,
so if an access to the cache happens during one of the scans,
it will block for a very small amount of time.

Additional improvements have been done to the internal helper APIs,
used to collect the users and groups information from Windows,
to avoid unnecessary data transformation round trips.